move schema type definitions from utils to types directory
